+++ src/CustomizeBuildings/NewComponents/SliderTemperatureSideScreen.cs
@@ -17,7 +17,7 @@ namespace CustomizeBuildings
         [Serialize] public float SetTemperature = 273.15f + 80f;
         [Serialize] public bool Switch;
 
-        public LocString TextLogic { get; set; } //Send <b><style="logic_on">Green Signal</style></b> if:
+        public LocString TextLogic; //Send <b><style="logic_on">Green Signal</style></b> if:
 
         public float Threshold { get => this.SetTemperature; set { this.SetTemperature = value; Update(); } }
         public bool ActivateAboveThreshold { get => this.Switch; set => this.Switch = value; }
@@ -30,10 +30,10 @@ namespace CustomizeBuildings
         public string BelowToolTip => STRINGS.UI.UISIDESCREENS.THRESHOLD_SWITCH_SIDESCREEN.BELOW_BUTTON;
         public ThresholdScreenLayoutType LayoutType => ThresholdScreenLayoutType.InputField;
         public int IncrementScale => 1;
-        public NonLinearSlider.Range[] GetRanges => NonLinearSlider.GetDefaultRange(((IThresholdSwitch)this).RangeMax);
+        public NonLinearSlider.Range[] GetRanges => NonLinearSlider.GetDefaultRange(this.RangeMax);
         public string Format(float value, bool units) => GameUtil.GetFormattedTemperature(value, GameUtil.TimeSlice.None, GameUtil.TemperatureInterpretation.Absolute, units, false);
-        public float GetRangeMaxInputField() => GameUtil.GetConvertedTemperature(((IThresholdSwitch)this).RangeMax, false);
-        public float GetRangeMinInputField() => GameUtil.GetConvertedTemperature(((IThresholdSwitch)this).RangeMin, false);
+        public float GetRangeMaxInputField() => GameUtil.GetConvertedTemperature(this.RangeMax, false);
+        public float GetRangeMinInputField() => GameUtil.GetConvertedTemperature(this.RangeMin, false);
         public float ProcessedInputValue(float input) => GameUtil.GetTemperatureConvertedToKelvin(input);
         public float ProcessedSliderValue(float input) => Mathf.Round(input);
         public LocString ThresholdValueUnits() => Helpers.GetTemperatureUnit();
@@ -58,7 +58,7 @@ namespace CustomizeBuildings
     /// <summary>
     /// Overwriting strings of hidden text fields.
     /// </summary>
-    [HarmonyPatch(typeof(ThresholdSwitchSideScreen), nameof(ThresholdSwitchSideScreen.OnSpawn))]
+    [HarmonyPatch(typeof(ThresholdSwitchSideScreen), nameof(ThresholdSwitchSideScreen.UpdateLabels))]
     public static class ThresholdSwitchSideScreen_FixTitle
     {
         public static void Postfix(ThresholdSwitchSideScreen __instance)
@@ -69,10 +69,12 @@ namespace CustomizeBuildings
                 var loc1 = default(LocString);
 
                 var slider = __instance.target?.GetComponent<SliderTemperatureSideScreen>();
-                if (slider != null)
-                {
+                if (slider != null)                
                     loc1 = slider.TextLogic;
-                }
+
+                var slider2 = __instance.target?.GetComponent<AirConditionerSliders>();
+                if (slider2 != null)
+                    loc1 = slider2.TextLogic;
 
                 locTexts[1].SetText(loc1 ?? STRINGS.UI.UISIDESCREENS.THRESHOLD_SWITCH_SIDESCREEN.ACTIVATE_IF);
             }