+++ src/CustomizeRecipe/_CustomizeRecipeState.cs
@@ -78,8 +78,8 @@ namespace CustomizeRecipe
         public static string GetStaticConfigPath()
         {
             string path = FumiKMod.ModName;
-            if (Helpers.ActiveLocale.NotEmpty() && Helpers.ActiveLocale != "en")
-                path += "_" + Helpers.ActiveLocale;
+            //if (Helpers.ActiveLocale.NotEmpty() && Helpers.ActiveLocale != "en")
+            //    path += "_" + Helpers.ActiveLocale;
             return Config.PathHelper.CreatePath(path);
         }