+++ src/CustomizeBuildings/Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.91.0] U51-600112-S
+- removed property translation #63 (sorry to reset some people's settings)
+- fixed HVAC compatibility
+- adjusted text on custom sliders
+
 ## [1.0.90.0] U51-600112-S
 - fix attempt on air conditioner